requirements.txt
asttokens==2.4.1
bleach==6.1.0
blinker==1.8.2
bokeh==3.3.4
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
executing==2.0.1
Flask==3.0.3
fonttools==4.53.1
idna==3.7
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
kiwisolver==1.4.5
linkify-it-py==2.0.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
numpy==1.26.4
packaging==24.1
pandas==2.2.2
panel==1.3.1
param==2.1.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
pydicom==2.4.4
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyvista==0.43.1
pyviz_comms==3.0.3
PyYAML==6.0.2
requests==2.32.3
scooby==0.10.0
six==1.16.0
stack-data==0.6.3
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==2.2.2
vtk==9.3.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
xyzservices==2024.6.0
